Fetch all companies. Har rydet op i bo.js, og lavet en crud.js til crud oprationer.
