+++ src/main/java/rest/ApplicationConfig.java
@@ -23,7 +23,6 @@ public class ApplicationConfig extends Application {
      */
     private void addRestResourceClasses(Set<Class<?>> resources) {
         resources.add(ExceptionHandling.PersonNotFoundExceptionMapper.class);
-      
         resources.add(rest.CompanyResource.class);
         resources.add(rest.PersonsResource.class);
     }
