+++ src/main/webapp/index.html
@@ -104,7 +104,7 @@
                             </div>
 
                             <div class='inputs form-group' id="btns">      
-                                <input id="userInput" placeholder='Add input' />
+                                <input id="userInputCrud" placeholder='Add input' />
                             </div>
 
                             <button class="btn btn-default" id="btncrud"> Submit </button>
@@ -407,6 +407,7 @@
         <!-- Custom scripts for this template -->
         <script src="bootstrap/js/freelancer.js" type="text/javascript"></script>
 
+        <script src="javascript/crud.js" type="text/javascript"></script>
         <script src="javascript/jsBenedikte.js" type="text/javascript"></script>
         <script src="javascript/jsBo.js" type="text/javascript"></script>
         <script src="javascript/jsChristian.js" type="text/javascript"></script>
